requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20
boto3==1.36.14
python-dotenv==1.0.1
supabase==2.14.0
Pillow==11.1.0
pydantic==2.10.6
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
httpx==0.27.2